requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.18.0
boto3==1.34.0
aws-requests-auth==0.4.3
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
pydantic==2.5.0
gunicorn==21.2.0